DataStructures

Documentation

type ProcCount = Int

counts number of times code has been evaluated against test cases

type LJMonad a = WriterT String (StateT ProcCount (ErrorT String IO)) a

Require logging, state and need to throw exceptions during computation

type TestCase = String

Content of the input file

type Answer = String

Content of the output file

data Problem

group together a problem atatement, its input and outputs and max allowed time limit

Constructors

Problem

Fields

problemStatement :: String

testCases :: [(TestCase, Answer)]

problemSetterCode :: String

timeLimit :: Int

data Contest

Contest is a list of problems

Constructors

Contest

Fields

problems :: [Problem]

data StatusCode

Status code determines the result of evaluation of the submitted code

Constructors

AC

WA

TLE

NZEC

type SubmissionResult = [(TestCase, StatusCode)]

data MuevalCommand

All that is required to run a mueval command

Constructors

MuevalCommand

Fields

expression :: String

testData :: String

ansData :: String

solutionFile :: String

maxRunTime :: Int