Lambdajudge

Description

Documentation

muevalAvlbl :: IO Bool

Check if mueval is installed on this system

gradeSubmission :: Problem -> FilePath -> LJMonad [StatusCode]

maps evaluation of solution file to each test case of the problem. Effectively we are sequencing the runs corresponding to each test case

runLJMonad :: Num s => WriterT w (StateT s (ErrorT e m)) a -> m (Either e ((a, w), s))

runs submitted code on problem testcases